refactor: dto 스타일의 객체 패키지 이동 및 상태 관련 로직 이동
