+++ src/main/java/bowling/domain/frame/state/Ready.java
@@ -2,7 +2,7 @@ package bowling.domain.frame.state;
 
 import java.util.List;
 
-import bowling.domain.BowlRecord;
+import bowling.domain.dto.BowlRecord;
 import bowling.domain.frame.Score;
 
 public class Ready extends Running {
@@ -18,7 +18,7 @@ public class Ready extends Running {
 
     @Override
     public BowlRecord createBowlRecord() {
-        return new BowlRecord(List.of());
+        return new BowlRecord(List.of(), false, false);
     }
 
     @Override
